Started trying to implement letrec so that I could make some of the vector functions in the stdlib. However, it turns out that my macro implementation is not good enough to be used for a letrec which is needed for a do/loop macro of somekind.
While figuring out what was wrong I implemented a
special form that can be used to expand a macro
expression once. This will be useful for debugging
and could be used as inspiration for a more full
implementation of a macro-expand function. I do
want to be able to take an arbitrairy expression
and expand it fully if it is or contains macros.
At the very least it means making it recursive so
that it maps over the arguments to the macro
somehow when it has been expanded. However, in
the state it is in it requires some evaluation to
have happened. Perhaps in the futrue we could
set it up so that we did a pass before evaluation
to load all macros and then did a pass where we
expanded all macros before evaluating. Or at least
made this kind of thing possible so that we can
do compilation after expanding macros.

Added a couple tests for vector additions, but
commented them out. I will uncomment them and
just implemet them in rust before testing them out
again.