Got basic map stuff working. Still missing the conversion and things like freeze.
For map I rewrote the reader to convert the map
literal syntax to (table ...) and (dict ...).
I am not really sure if this is the right thing
to do or not. It makes ensuring evaluation of
arguments easier, but I think it removes the
ability to read them in other contexts. I tried
changing vector, but its use as structure in things
like lambda made it break everything. What I worry
about is that evaluating arguments to literal
structures in vm could end up evaluating their
members more than once. I.e. I read a map as a
function argument and evaluate it and its contents,
then return that map to be used somewhere else and
it gets evaluated a second time. Though I expect the
only way to pass it around is to pass it indirectly
through a symbol or as a return. So, in those cases
the form being evaluated is the symbol or function
and not the map. But if a map or vector had a list
as a value evaluating it again would be an error
if the list didn't represent a valid function call.
Though I suppose that any list that is created through
(list ...) is passed around as a list and lists
are all evaluated whenever they are seen directly
and that is not a problem.